Extract loading_indicator.rs - a reusable function
